+++ src/app/dorasena/page.tsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import Dorasena01 from "./Dorasena01/dorasena";
+import Dorasena01 from "../../app/dorasena";
 import Image from "next/image";
 
 export default function Dorasena() {
